refactor(ai): split expand_phase_planner into part-of concern files (#3278)
Item 5 maintainability split: extract the 2,626-physical-line
expand_phase_planner.dart (well over the 1000-line code-review
guideline) into single-concern `part of` files. The main library drops
to 778 lines; no concern file exceeds ~575 lines.

New part files (same library scope — no import/visibility changes):
- expand_phase_planner_economy.dart (ExpandEconomyPlan, planExpandEconomy,
  below-quota treasury-recovery predicates)
- expand_phase_planner_military.dart (ExpandMilitaryPlan, planExpandMilitary,
  same-turn declare-war coordination helpers)
- expand_phase_planner_peace_targets.dart (sole-GP / quota / critical-hold
  peace deciders)
- expand_phase_planner_peace_default_start.dart (default-start / near-quota /
  distraction peace family + stalledOwExpansionNeedsPeacePass composite)

Behaviour-preserving: code moved byte-identically; main diff is only the
four new part directives plus deletions. Determinism unchanged.

Refs #3278
